\docType{data}
\name{abu}
\alias{abu}
\title{Species sampled at eight sites around an outfall.
Site locations are designated as kilometers upstream (-) or downstream (+) of the outfall.}
\format{A data.frame with 8 observations of abundance of 10 Species.}
\usage{
data(abu)
}
\description{
Species sampled at eight sites around an outfall. Site
locations are designated as kilometers upstream (-) or
downstream (+) of the outfall.
}
\keyword{data}

