\docType{data}
\name{napcp}
\alias{napcp}
\title{Fathead minnow (\emph{Pimephales promelas}) larval survival data for sodium pentachlorophenol (NaPCP) exposure}
\format{A data.frame with 2 columns:
\describe{
\item{"arc"}{numeric; Proportion of fish alive at the end of experiment. Proportions have been arcsine square root transformed.}
\item{"conc"}{factor; NaPCP concentration.}
}}
\usage{
data(napcp)
}
\description{
Fathead minnow (\emph{Pimephales promelas}) larval survival
data for sodium pentachlorophenol (NaPCP) exposure
}
\note{
Proportions have been 1.5708 - arcsin(sqrt(1/(4*x)))
transformed.
}
\references{
Weber et al. 1989. Short-term methods for estimating the
chronic toxicity of effluents and receiving waters to
freshwater organisms. EPA.
}
\keyword{data}

