\docType{data}
\name{naphta}
\alias{naphta}
\title{Mortality of grass shrimp at various naphthalene concentrations}
\format{A data.frame with 3 columns:
\describe{
\item{"conc"}{Naphtalene concentration.}
\item{"dead"}{Number of dead shrimp}
\item{"total"}{Total number of shrimp exposed.}
}}
\usage{
data(naphta)
}
\description{
In a study of the lethal effects of PAH compounds released
during oil spills, Unger et al. (2008) exposed grass shrimp
(Palaemonetes pugio) to a range of naphthalene
concentrations.
}
\references{
Unger, M.A., M.C. Newman, and G.G. Vadas. Predicting
survival of grass shrimp, Palamonetes pugio, exposed to
naphthalene, fluorine, and dibenzothiophene. Environ.
Toxicol. Chem. 27(8):1802–1808 (2008).
}
\keyword{data}

