\docType{data}
\name{oil}
\alias{oil}
\alias{oil_env}
\title{Benthic community data around an offshore installation}
\format{A dataframe of observations of 50 species at 20 sites.}
\source{
Data was retrieved from the UK Benthos Database,Oil & Gas
UK ,
http://www.oilandgasuk.co.uk/knowledgecentre/uk_benthos_database.cfm

Data was retrieved from the UK Benthos Database,Oil & Gas
UK ,
http://www.oilandgasuk.co.uk/knowledgecentre/uk_benthos_database.cfm
}
\usage{
data(oil)

data(oil_env)
}
\description{
Samples were collected at sampling stations with varying
distanced from installation along the transects. For
environmental (and distance) data see
\code{\link{oil_env}}.
}
\seealso{
\code{\link{oil_env}}

\code{\link{oil}}
}
\keyword{data}

