\docType{data}
\name{oil_env}
\alias{oil_env}
\title{Corresponding environmental data to oil data set}
\format{A dataframe of 20 observations of 22 variables:
\describe{
\item{distance}{distance from installation in m.}
\item{site}{A sampling site identifier.}
\item{silt_clay}{Silt/clay content of the sediment as percent.}
\item{organic}{Organic content of the sedimentas percent.}
\item{tot_hc_gc}{Total Hydrocarbon determined by gas chromatography.}
\item{naph_ult}{Naphthalenes (via ultrasonification).}
\item{phen_ult}{Phenanthrenes (via ultrasonification).}
\item{dibenz_ult}{Diabenzothiophenes (via ultrasonification).}
\item{pah_4r_202}{Four ring compounds (mass 202).}
\item{pah_4r_228}{Four ring compounds (mass 228).}
\item{pah_5r_252}{Five ring compounds (mass 252).}
\item{pah_6r_276}{Six ring compounds (mass 2575).}
\item{ba_tot}{Total Barium concentration.}
\item{cd}{Cadmium.}
\item{cr}{Chromium.}
\item{cu}{Copper.}
\item{ni}{Nickel.}
\item{pb}{Lead.}
\item{va}{Vanadium.}
\item{zn}{Zinc.}
\item{fe}{Iron.}
\item{hg}{mercury}
}}
\source{
  Data was retrieved from the UK Benthos Database,Oil & Gas
  UK ,
  http://www.oilandgasuk.co.uk/knowledgecentre/uk_benthos_database.cfm
}
\description{
  Corresponding environmental data to oil data set
}
\seealso{
  \code{\link{oil}}
}
\keyword{data}
\keyword{datasets}

