\docType{data}
\name{salt}
\alias{salt}
\title{Dead of mosquitofish at different NaCl concentrations}
\format{A data.frame with 3 columns:
\describe{
\item{"dead"}{Number of fish died}
\item{"total"}{Total number of fish exposed.}
\item{"conc"}{NaCl Concentration.}
}}
\description{
  Newman and Aplin (1992) exposed mosquitofish
  (\emph{Gambusia holbrooki}) to a series of NaCl
  concentrations for 96h.
}
\references{
  Newman, M.C. and M. Aplin.  1992.  Enhancing toxicity
  data interpretation and prediction of ecological risk
  with survival time modeling: An illustration using sodium
  chloride toxicity to mosquitofish (\emph{Gambusia
  holbrooki}).  Aquatic Toxicol. 23:85-96.
}
\keyword{data}
\keyword{datasets}

