\docType{data}
\name{oil}
\alias{oil}
\title{Benthic community data around a offshore installation}
\format{A dataframe of observations of 50 species at 20 sites.}
\source{
  Data was retrieved from the UK Benthos Database,Oil & Gas
  UK ,
  http://www.oilandgasuk.co.uk/knowledgecentre/uk_benthos_database.cfm
}
\description{
  Samples were collected at sampling stations with varying
  distanced from installation along the transects. For
  environmental (and distance) data see
  \code{\link{oil_env}}.
}
\seealso{
  \code{\link{oil_env}}
}
\keyword{data}
\keyword{datasets}

