\docType{data}
\name{so4}
\alias{so4}
\title{Sulfate concentrations (mg/L) measured during a routine water quality survey
of the Savannah River (South Carolina).}
\value{
  A data.frame with the variable \emph{conc}.
}
\description{
  Sulfate concentrations (mg/L) measured during a routine
  water quality survey of the Savannah River (South
  Carolina).
}
\keyword{data}
\keyword{datasets}

