\docType{data}
\name{so4}
\alias{so4}
\title{Sulfate concentrations (mg/L) measured during a routine water quality survey
of the Savannah River (South Carolina).}
\format{A vector with 21 measurements.}
\usage{
data(so4)
}
\description{
Sulfate concentrations (mg/L) measured during a routine
water quality survey of the Savannah River (South
Carolina).
}
\keyword{data}

